requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-docx
pywin32==303
